Moves toJson function into user.coffee.
Fixe #67

The underlying problem in #67 is that `toJson` isn't defined when the oauth response is pulled out of the cache.  Rather than having to always ensure this method exists on the oauthResponse object (which can come from multiple places), it makes more sense to just make that function accessible from anywhere in `user.coffee`.